Split out NumberRollingColumnTest from RollingColumnTest
